// src/utils/types/mapped-event-project.ts
export type MappedEventProject = {
  name: string;
  distance: string;
  elevation: string;
  time: string;
  companionship: string;
  islands?: string;
  munros?: string;
  munroTops?: string;
  corbetts?: string;
  corbettTops?: string;
  grahams?: string;
  subTwos?: string;
  donalds?: string;
};